Fix up integer constants. Test simple expressions.

+++ src/ast.rs
@@ -54,7 +54,7 @@ pub enum Expr
     True,
     False,
     Nil,
-    Int(i128),
+    Int64(i64),
     Float64(f64),
     String(String),
